bp: Move MaxPreferredWorkGroupInvocations
